Tasks 1/2

+++ runtime/runtime_lib.cc
@@ -1,6 +1,6 @@
 #include <cstdio>
 
 extern "C"
-void printi(int i) {
-    printf("%d\n", i);
+void printi(long i) {
+    printf("%ld\n", i);
 }
